i fuckin hate asyncio

+++ src/robot_map.py
@@ -3,7 +3,7 @@ import numpy as np
 
 class RobotMap:
     def __init__(self):
-        self.map = np.zeros((640, 480, 3))
+        self.map = np.zeros((640, 480, 3), dtype=np.uint8)
         self.instruction_data = []
 
     def draw_interest_point(self, X, Y):
